requirements.txt
# NeuronOS 项目依赖
# 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 异步支持
aiofiles==23.2.1
aiohttp==3.9.1

# 数据库
neo4j==5.15.0
aiosqlite==0.19.0

# 日志处理
loguru==0.7.2

# 配置管理
python-dotenv==1.0.0
PyYAML==6.0.1

# 测试框架
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# AI/ML 依赖
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2

# 向量数据库和嵌入
pinecone-client==3.0.0
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
tiktoken==0.5.2

# 其他工具
requests==2.31.0
psutil==5.9.6
watchdog==3.0.0